Extracting shared `Action` component to simplify `Field`/`ValueActions`.
