--- graylog2-web-interface/src/views/components/Value.css
@@ -1,16 +1,0 @@
-:local(.dropdownheader) {
-    padding-left: 10px;
-    padding-right: 10px;
-    padding-bottom: 5px;
-    margin-bottom: 5px;
-    font-weight: 600;
-}
-
-:local(.bottomSpacer) {
-    margin-bottom: 10px;
-    list-style: none;
-}
-
-:local(.topSpacer) {
-    margin-top: 10px;
-}
